Final data is now sliced to remove the bump and slowdown in the pavilion.
I added a new signal type 'taskSignals' which are truncated versions of the
computedSignals. I remove the bump portion and any data taken after the end of
the track in the pavilion. This should result in cleaner data sets for the
system identification. I also added the ability to scale run plot signals
individually for better plotting readability.